Progress commit on attempting to set up Jest for testing. Hit some road blocks along the way as it needs to be configured into webpack to function correctly. The data model classes are also not set up well to work as loosly coupled objects for testing. Some work has been started reorganize how those object get built.
